+++ bip32/index.js
@@ -1,5 +1,6 @@
 const bip32Derivation = require('./bip32_derivation');
 const bip32Path = require('./bip32_path');
+const encodeDerivations = require('./encode_derivations');
 const taprootBip32 = require('./taproot_bip32');
 
-module.exports = {bip32Derivation, bip32Path, taprootBip32};
+module.exports = {bip32Derivation, bip32Path, encodeDerivations, taprootBip32};