+++ index.js
@@ -2,6 +2,7 @@ const {combinePsbts} = require('./psbts');
 const {createPsbt} = require('./psbts');
 const {decodePsbt} = require('./psbts');
 const {encodePsbt} = require('./psbts');
+const {extendPsbt} = require('./psbts');
 const {extractTransaction} = require('./psbts');
 const {finalizePsbt} = require('./psbts');
 const {signPsbt} = require('./psbts');
@@ -13,6 +14,7 @@ module.exports = {
   createPsbt,
   decodePsbt,
   encodePsbt,
+  extendPsbt,
   extractTransaction,
   finalizePsbt,
   signPsbt,